Move check_device_ids down into devices.rs
Even though it's temporary, it clearly belongs there, as it handles the
DeviceInfos which are generated by process_devices.

Use check_device_ids in tests in devices.rs where applicable.

Define a new method called process_and_verify_devices.
Use it to combine the functionality of process_devices and
check_device_ids.

Revise a stale post-condition from check_device_ids header comment.
It became stale in a previous commit, when DeviceInfo was changed so
that it no longer contained a field for Stratis identifiers.
However, the general outlines of the post-condition remain the same.
